reimplement info using typescript/esm
