+++ generators/info/generator.spec.mts
