+++ generators/info/index.mts
@@ -16,4 +16,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default } from './generator.cjs';
+export { default } from './generator.mjs';
